Change all occurences of 'video' to 'lesson' in VideoListHeader storybook file
